Resolve many asan runtime errors (#5586)
The small set of fixes in this PR resolve almost 900 errors
reported by asan against the BDE test drivers.  These are
all errors in the test drivers, and are so numerous that
they drown out any reporting of other issues.  They follow
the same pattern that an object is created in an object
buffer and then moved and never destroyed.  The reported
leaks occur when a test type is stored that owns memory in
its moved-from state.

There is one additional patch to resolve an array new that
is never deleted during a breathing test, so that the whole
test driver can run clean.  The fix is to use an array that
is a local variable rather than dynamically allocated."