Fix memory leaks in `balcl_commandline` tests reported by asan (#5602)
There are several source of leaks and memory corruption in this test
driver, as reported by asan.  Installing a test allocator as default
in `main` goes a long way to identifying the issues.

First, we observe that a lot of global data is locking down the
default allocator during global initialization, before `main`.
We try to address that by supplying a specific allocator for
each global, but we cannot reach the globals in our test table
arrays, as you cannot supply an allocator to an array.

After doing our best to provide allocators to all the global
variables we force a test allocator as the default in `main`
using a default allocator guard, which bypasses the lock.
This change reveals a signicant concern where function local
statics are being destroyed after `main`, so after the
default allocator has been destroyed.  We resolve those
issues by making the test tables regular local variables.
rather than static, in those functions.