DRQS 182144710 Remove `using namespace std;` (#5694)
* DRQS 182144710 Remove using namespace std;

* Remove more unneeded using namespace std

The test drivers include only the C standard library headers, so the
using namespace std to pull the stnadard library functions from namespace
std into the global namespace is not only unnecessary, but also might not
find the namespace at all.  In these cases, there is no warning as the
the other 'bsl' headers included happen to transitively include a header
that opens namespace std, such as <type_traits>, but that is relying on
a transitive detail that is strictly unnecessary for the code under test.